Revert "Bugfix #9673: Origin header is always sent from WebSocket client (#9692)"
This reverts commit f48d9fa8d05cfe5848509d0ed2739cc857d071ce as this needs more thoughts.
